Added New Creation of Tests and avoided printing > 100 lists

+++ number_list_loader.h
@@ -7,7 +7,7 @@ struct number_list{
     int size;
     std::string test_name;
     int k;
-    int expected_result;
+    double long expected_result;
     double* nubmers;
 };
